[UPDATE] JPA Controller

+++ src/java/controller/AccountJpaController.java
@@ -5,17 +5,21 @@
  */
 package controller;
 
+import controller.exceptions.IllegalOrphanException;
 import controller.exceptions.NonexistentEntityException;
 import controller.exceptions.PreexistingEntityException;
 import controller.exceptions.RollbackFailureException;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import model.Customer;
+import model.History;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 import javax.transaction.UserTransaction;
 import model.Account;
 
@@ -37,11 +41,43 @@ public class AccountJpaController implements Serializable {
     }
 
     public void create(Account account) throws PreexistingEntityException, RollbackFailureException, Exception {
+        if (account.getHistoryList() == null) {
+            account.setHistoryList(new ArrayList<History>());
+        }
         EntityManager em = null;
         try {
             utx.begin();
             em = getEntityManager();
+            Customer customer = account.getCustomer();
+            if (customer != null) {
+                customer = em.getReference(customer.getClass(), customer.getEmail());
+                account.setCustomer(customer);
+            }
+            List<History> attachedHistoryList = new ArrayList<History>();
+            for (History historyListHistoryToAttach : account.getHistoryList()) {
+                historyListHistoryToAttach = em.getReference(historyListHistoryToAttach.getClass(), historyListHistoryToAttach.getHistoryid());
+                attachedHistoryList.add(historyListHistoryToAttach);
+            }
+            account.setHistoryList(attachedHistoryList);
             em.persist(account);
+            if (customer != null) {
+                Account oldAccountOfCustomer = customer.getAccount();
+                if (oldAccountOfCustomer != null) {
+                    oldAccountOfCustomer.setCustomer(null);
+                    oldAccountOfCustomer = em.merge(oldAccountOfCustomer);
+                }
+                customer.setAccount(account);
+                customer = em.merge(customer);
+            }
+            for (History historyListHistory : account.getHistoryList()) {
+                Account oldEmailOfHistoryListHistory = historyListHistory.getEmail();
+                historyListHistory.setEmail(account);
+                historyListHistory = em.merge(historyListHistory);
+                if (oldEmailOfHistoryListHistory != null) {
+                    oldEmailOfHistoryListHistory.getHistoryList().remove(historyListHistory);
+                    oldEmailOfHistoryListHistory = em.merge(oldEmailOfHistoryListHistory);
+                }
+            }
             utx.commit();
         } catch (Exception ex) {
             try {
@@ -60,12 +96,66 @@ public class AccountJpaController implements Serializable {
         }
     }
 
-    public void edit(Account account) throws NonexistentEntityException, RollbackFailureException, Exception {
+    public void edit(Account account) throws IllegalOrphanException, NonexistentEntityException, RollbackFailureException, Exception {
         EntityManager em = null;
         try {
             utx.begin();
             em = getEntityManager();
+            Account persistentAccount = em.find(Account.class, account.getEmail());
+            Customer customerOld = persistentAccount.getCustomer();
+            Customer customerNew = account.getCustomer();
+            List<History> historyListOld = persistentAccount.getHistoryList();
+            List<History> historyListNew = account.getHistoryList();
+            List<String> illegalOrphanMessages = null;
+            if (customerOld != null && !customerOld.equals(customerNew)) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("You must retain Customer " + customerOld + " since its account field is not nullable.");
+            }
+            for (History historyListOldHistory : historyListOld) {
+                if (!historyListNew.contains(historyListOldHistory)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain History " + historyListOldHistory + " since its email field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            if (customerNew != null) {
+                customerNew = em.getReference(customerNew.getClass(), customerNew.getEmail());
+                account.setCustomer(customerNew);
+            }
+            List<History> attachedHistoryListNew = new ArrayList<History>();
+            for (History historyListNewHistoryToAttach : historyListNew) {
+                historyListNewHistoryToAttach = em.getReference(historyListNewHistoryToAttach.getClass(), historyListNewHistoryToAttach.getHistoryid());
+                attachedHistoryListNew.add(historyListNewHistoryToAttach);
+            }
+            historyListNew = attachedHistoryListNew;
+            account.setHistoryList(historyListNew);
             account = em.merge(account);
+            if (customerNew != null && !customerNew.equals(customerOld)) {
+                Account oldAccountOfCustomer = customerNew.getAccount();
+                if (oldAccountOfCustomer != null) {
+                    oldAccountOfCustomer.setCustomer(null);
+                    oldAccountOfCustomer = em.merge(oldAccountOfCustomer);
+                }
+                customerNew.setAccount(account);
+                customerNew = em.merge(customerNew);
+            }
+            for (History historyListNewHistory : historyListNew) {
+                if (!historyListOld.contains(historyListNewHistory)) {
+                    Account oldEmailOfHistoryListNewHistory = historyListNewHistory.getEmail();
+                    historyListNewHistory.setEmail(account);
+                    historyListNewHistory = em.merge(historyListNewHistory);
+                    if (oldEmailOfHistoryListNewHistory != null && !oldEmailOfHistoryListNewHistory.equals(account)) {
+                        oldEmailOfHistoryListNewHistory.getHistoryList().remove(historyListNewHistory);
+                        oldEmailOfHistoryListNewHistory = em.merge(oldEmailOfHistoryListNewHistory);
+                    }
+                }
+            }
             utx.commit();
         } catch (Exception ex) {
             try {
@@ -88,7 +178,7 @@ public class AccountJpaController implements Serializable {
         }
     }
 
-    public void destroy(String id) throws NonexistentEntityException, RollbackFailureException, Exception {
+    public void destroy(String id) throws IllegalOrphanException, NonexistentEntityException, RollbackFailureException, Exception {
         EntityManager em = null;
         try {
             utx.begin();
@@ -99,6 +189,24 @@ public class AccountJpaController implements Serializable {
                 account.getEmail();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The account with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            Customer customerOrphanCheck = account.getCustomer();
+            if (customerOrphanCheck != null) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Account (" + account + ") cannot be destroyed since the Customer " + customerOrphanCheck + " in its customer field has a non-nullable account field.");
+            }
+            List<History> historyListOrphanCheck = account.getHistoryList();
+            for (History historyListOrphanCheckHistory : historyListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Account (" + account + ") cannot be destroyed since the History " + historyListOrphanCheckHistory + " in its historyList field has a non-nullable email field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(account);
             utx.commit();
